cleaned up functions
