+++ services/app-api/handlers/banners/get.ts
@@ -2,7 +2,7 @@ import handler from "../handler-lib";
 import dynamoDb from "../../utils/dynamo/dynamodb-lib";
 import { StatusCodes } from "../../utils/types/types";
 import { NO_KEY_ERROR_MESSAGE } from "../../utils/constants/constants";
-import { sanitize } from "../../utils/sanitize";
+import { sanitizeObject } from "../../utils/sanitizeObject";
 
 export const getBanner = handler(async (event, _context) => {
   if (!event?.pathParameters?.bannerId!) {
@@ -17,9 +17,7 @@ export const getBanner = handler(async (event, _context) => {
   };
   const queryResponse: any = await dynamoDb.get(params);
 
-  queryResponse.Item.title = sanitize(queryResponse.Item.title);
-  queryResponse.Item.description = sanitize(queryResponse.Item.description);
-  queryResponse.Item.link = sanitize(queryResponse.Item.link);
+  queryResponse.Item = sanitizeObject(queryResponse.Item);
 
   return { status: StatusCodes.SUCCESS, body: queryResponse };
 });
